Sema: Split off TypeOfReference.cpp from ConstraintSystem.cpp
